tests/wpt/web-platform-tests/shadow-dom/untriaged/events/event-retargeting/test-002.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_05_01_02</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#event-retargeting">
<meta name="assert" content="Event Retargeting:Event retargeting for document nodes distributed among insertion points">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
var A_05_01_02_T1 = async_test('A_05_01_02_T1');

A_05_01_02_T1.step(function () {
    var iframe = document.createElement('iframe');
    iframe.src = '../../resources/bobs_page.html';
    document.body.appendChild(iframe);

    iframe.onload = A_05_01_02_T1.step_func(function () {

    try {
    	var d = iframe.contentDocument;

        var ul = d.querySelector('ul.stories');
        var s = ul.createShadowRoot();

        //make shadow subtree
        var div = document.createElement('div');
        div.innerHTML = '<ul id="ip_wrapper"><content select=".shadow"></content></ul>';
        s.appendChild(div);

    	d.body.addEventListener('click', A_05_01_02_T1.step_func(function (event) {
            assert_equals(event.target.tagName, 'LI', 'Incorrect target');
        }), false);

        var event = d.createEvent('HTMLEvents');
        event.initEvent ("click", true, false);
        d.querySelector('#li3').dispatchEvent(event);
    } finally {
        iframe.parentNode.removeChild(iframe);
    }
    A_05_01_02_T1.done();
    });
});
</script>
</body>
</html>
